# eval/cases-ts/TN02/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fixture Payments API (TypeScript)",
    "version": "1.0.0",
    "description": "Synthetic payments API used as an evaluation target for the contract auditor."
  },
  "servers": [
    {
      "url": "https://api.fixture.test/v1"
    }
  ],
  "paths": {
    "/payouts": {
      "post": {
        "tags": [
          "Payouts"
        ],
        "summary": "Create a payout",
        "description": "Initiates a payout to a bank account. Requires an Idempotency-Key header.",
        "operationId": "createPayout",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Duplicate reference"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Unique key to make the request idempotent",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayoutRequest"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/payouts/{id}": {
      "get": {
        "tags": [
          "Payouts"
        ],
        "summary": "Retrieve a payout",
        "description": "Fetches a single payout by its identifier.",
        "operationId": "getPayout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Payout ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/refunds": {
      "post": {
        "tags": [
          "Payouts"
        ],
        "summary": "Create a refund",
        "description": "Refunds a previously successful payout, in full or in part.",
        "operationId": "createRefund",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Refund not permitted"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/balance": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Retrieve wallet balance",
        "description": "Returns the available and ledger balance for the authenticated account.",
        "operationId": "getBalance",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "List transactions",
        "description": "Returns a paginated list of transactions, newest first.",
        "operationId": "listTransactions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number, 1-indexed",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "description": "Results per page, max 100",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a customer",
        "description": "Creates a customer record.",
        "operationId": "createCustomer",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRequest"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/kyc/bvn": {
      "post": {
        "tags": [
          "KYC"
        ],
        "summary": "Verify a BVN",
        "description": "Verifies a Bank Verification Number and returns the matched identity.",
        "operationId": "verifyBVN",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BVNResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BVNRequest"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/banks": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "List supported banks",
        "description": "Returns every institution that can receive a payout.",
        "operationId": "listBanks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bank"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Send a test webhook",
        "description": "Delivers a sample event to the configured webhook URL. The delivery is signed with HMAC-SHA256 in the X-Signature header.",
        "operationId": "testWebhook",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "PayoutRequest": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string"
          },
          "bankCode": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "description": "Decimal string in minor units to avoid float drift"
          },
          "currency": {
            "type": "string"
          },
          "narration": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          }
        },
        "required": [
          "accountNumber",
          "bankCode",
          "amount",
          "currency",
          "reference"
        ]
      },
      "PayoutResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "available": {
            "type": "string"
          },
          "ledger": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "CustomerRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "kycTier": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransactionsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PayoutResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "perPage": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "BVNRequest": {
        "type": "object",
        "properties": {
          "bvn": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "bvn"
        ]
      },
      "BVNResponse": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "tier": {
            "type": "integer"
          }
        }
      },
      "RefundRequest": {
        "type": "object",
        "properties": {
          "payoutId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "payoutId",
          "amount"
        ]
      },
      "RefundResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "payoutId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          }
        }
      },
      "Bank": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  }
}